server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "joblib>=1.5.3",
    "openai>=2.14.0",
    "passlib[bcrypt]>=1.7.4",
    "pinecone-client>=6.0.0",
    "prisma>=0.15.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.40.0",
]
